@@ -20,16 +20,17 @@ class SharedPreferencesLinux extends SharedPreferencesStorePlatform {
   static SharedPreferencesLinux instance = SharedPreferencesLinux();
 
   /// Local copy of preferences
-  Map<String, Object> _cachedPreferences;
+  Map<String, Object>? _cachedPreferences;
 
   /// File system used to store to disk. Exposed for testing only.
   @visibleForTesting
   FileSystem fs = LocalFileSystem();
 
   /// Gets the file where the preferences are stored.
-  Future<File> _getLocalDataFile() async {
+  Future<File?> _getLocalDataFile() async {
     final pathProvider = PathProviderLinux();
     final directory = await pathProvider.getApplicationSupportPath();
+    if (directory == null) return null;
     return fs.file(path.join(directory, 'shared_preferences.json'));
   }
 
@@ -37,19 +38,19 @@ class SharedPreferencesLinux extends SharedPreferencesStorePlatform {
   /// maintained in memory.
   Future<Map<String, Object>> _readPreferences() async {
     if (_cachedPreferences != null) {
-      return _cachedPreferences;
+      return _cachedPreferences!;
     }
 
-    _cachedPreferences = {};
-    var localDataFile = await _getLocalDataFile();
-    if (localDataFile.existsSync()) {
+    Map<String, Object> preferences = {};
+    final File? localDataFile = await _getLocalDataFile();
+    if (localDataFile != null && localDataFile.existsSync()) {
       String stringMap = localDataFile.readAsStringSync();
       if (stringMap.isNotEmpty) {
-        _cachedPreferences = json.decode(stringMap) as Map<String, Object>;
+        preferences = json.decode(stringMap).cast<String, Object>();
       }
     }
-
-    return _cachedPreferences;
+    _cachedPreferences = preferences;
+    return preferences;
   }
 
   /// Writes the cached preferences to disk. Returns [true] if the operation
@@ -57,6 +58,10 @@ class SharedPreferencesLinux extends SharedPreferencesStorePlatform {
   Future<bool> _writePreferences(Map<String, Object> preferences) async {
     try {
       var localDataFile = await _getLocalDataFile();
+      if (localDataFile == null) {
+        print("Unable to determine where to write preferences.");
+        return false;
+      }
       if (!localDataFile.existsSync()) {
         localDataFile.createSync(recursive: true);
       }
