@@ -14,8 +14,7 @@ class _MyAppState extends State<MyApp> {
   void initState() {
     super.initState();
 
-    IosPlatformImages.resolveURL("textfile", null)
-        .then((value) => print(value));
+    IosPlatformImages.resolveURL("textfile").then((value) => print(value));
   }
 
   @override
