@@ -121,38 +121,50 @@ GyroscopeEvent _listToGyroscopeEvent(List<double> list) {
   return GyroscopeEvent(list[0], list[1], list[2]);
 }
 
-Stream<AccelerometerEvent> _accelerometerEvents;
-Stream<GyroscopeEvent> _gyroscopeEvents;
-Stream<UserAccelerometerEvent> _userAccelerometerEvents;
+Stream<AccelerometerEvent>? _accelerometerEvents;
+Stream<GyroscopeEvent>? _gyroscopeEvents;
+Stream<UserAccelerometerEvent>? _userAccelerometerEvents;
 
 /// A broadcast stream of events from the device accelerometer.
 Stream<AccelerometerEvent> get accelerometerEvents {
-  if (_accelerometerEvents == null) {
-    _accelerometerEvents = _accelerometerEventChannel
-        .receiveBroadcastStream()
-        .map(
-            (dynamic event) => _listToAccelerometerEvent(event.cast<double>()));
+  Stream<AccelerometerEvent>? accelerometerEvents = _accelerometerEvents;
+  if (accelerometerEvents == null) {
+    accelerometerEvents =
+        _accelerometerEventChannel.receiveBroadcastStream().map(
+              (dynamic event) =>
+                  _listToAccelerometerEvent(event.cast<double>()),
+            );
+    _accelerometerEvents = accelerometerEvents;
   }
-  return _accelerometerEvents;
+
+  return accelerometerEvents;
 }
 
 /// A broadcast stream of events from the device gyroscope.
 Stream<GyroscopeEvent> get gyroscopeEvents {
-  if (_gyroscopeEvents == null) {
-    _gyroscopeEvents = _gyroscopeEventChannel
-        .receiveBroadcastStream()
-        .map((dynamic event) => _listToGyroscopeEvent(event.cast<double>()));
+  Stream<GyroscopeEvent>? gyroscopeEvents = _gyroscopeEvents;
+  if (gyroscopeEvents == null) {
+    gyroscopeEvents = _gyroscopeEventChannel.receiveBroadcastStream().map(
+          (dynamic event) => _listToGyroscopeEvent(event.cast<double>()),
+        );
+    _gyroscopeEvents = gyroscopeEvents;
   }
-  return _gyroscopeEvents;
+
+  return gyroscopeEvents;
 }
 
 /// Events from the device accelerometer with gravity removed.
 Stream<UserAccelerometerEvent> get userAccelerometerEvents {
-  if (_userAccelerometerEvents == null) {
-    _userAccelerometerEvents = _userAccelerometerEventChannel
-        .receiveBroadcastStream()
-        .map((dynamic event) =>
-            _listToUserAccelerometerEvent(event.cast<double>()));
+  Stream<UserAccelerometerEvent>? userAccelerometerEvents =
+      _userAccelerometerEvents;
+  if (userAccelerometerEvents == null) {
+    userAccelerometerEvents =
+        _userAccelerometerEventChannel.receiveBroadcastStream().map(
+              (dynamic event) =>
+                  _listToUserAccelerometerEvent(event.cast<double>()),
+            );
+    _userAccelerometerEvents = userAccelerometerEvents;
   }
-  return _userAccelerometerEvents;
+
+  return userAccelerometerEvents;
 }
