@@ -52,16 +52,16 @@ void main() {
 void _initializeFakeSensorChannel(String channelName, List<double> sensorData) {
   const StandardMethodCodec standardMethod = StandardMethodCodec();
 
-  void _emitEvent(ByteData event) {
-    ServicesBinding.instance.defaultBinaryMessenger.handlePlatformMessage(
+  void _emitEvent(ByteData? event) {
+    ServicesBinding.instance!.defaultBinaryMessenger.handlePlatformMessage(
       channelName,
       event,
-      (ByteData reply) {},
+      (ByteData? reply) {},
     );
   }
 
-  ServicesBinding.instance.defaultBinaryMessenger
-      .setMockMessageHandler(channelName, (ByteData message) async {
+  ServicesBinding.instance!.defaultBinaryMessenger
+      .setMockMessageHandler(channelName, (ByteData? message) async {
     final MethodCall methodCall = standardMethod.decodeMethodCall(message);
     if (methodCall.method == 'listen') {
       _emitEvent(standardMethod.encodeSuccessEnvelope(sensorData));
