@@ -1,12 +1,13 @@
 import 'dart:async';
+
 import 'package:flutter_test/flutter_test.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 import 'package:integration_test/integration_test.dart';
+import 'package:shared_preferences_linux/shared_preferences_linux.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
-  group('$SharedPreferences', () {
+  group('SharedPreferencesLinux', () {
     const Map<String, dynamic> kTestValues = <String, dynamic>{
       'flutter.String': 'hello world',
       'flutter.bool': true,
@@ -23,10 +24,10 @@ void main() {
       'flutter.List': <String>['baz', 'quox'],
     };
 
-    SharedPreferences preferences;
+    SharedPreferencesLinux preferences;
 
     setUp(() async {
-      preferences = await SharedPreferences.getInstance();
+      preferences = SharedPreferencesLinux.instance;
     });
 
     tearDown(() {
@@ -34,56 +35,62 @@ void main() {
     });
 
     testWidgets('reading', (WidgetTester _) async {
-      expect(preferences.get('String'), isNull);
-      expect(preferences.get('bool'), isNull);
-      expect(preferences.get('int'), isNull);
-      expect(preferences.get('double'), isNull);
-      expect(preferences.get('List'), isNull);
-      expect(preferences.getString('String'), isNull);
-      expect(preferences.getBool('bool'), isNull);
-      expect(preferences.getInt('int'), isNull);
-      expect(preferences.getDouble('double'), isNull);
-      expect(preferences.getStringList('List'), isNull);
+      final all = await preferences.getAll();
+      expect(all['String'], isNull);
+      expect(all['bool'], isNull);
+      expect(all['int'], isNull);
+      expect(all['double'], isNull);
+      expect(all['List'], isNull);
     });
 
     testWidgets('writing', (WidgetTester _) async {
       await Future.wait(<Future<bool>>[
-        preferences.setString('String', kTestValues2['flutter.String']),
-        preferences.setBool('bool', kTestValues2['flutter.bool']),
-        preferences.setInt('int', kTestValues2['flutter.int']),
-        preferences.setDouble('double', kTestValues2['flutter.double']),
-        preferences.setStringList('List', kTestValues2['flutter.List'])
+        preferences.setValue(
+            'String', 'String', kTestValues2['flutter.String']),
+        preferences.setValue('Bool', 'bool', kTestValues2['flutter.bool']),
+        preferences.setValue('Int', 'int', kTestValues2['flutter.int']),
+        preferences.setValue(
+            'Double', 'double', kTestValues2['flutter.double']),
+        preferences.setValue('StringList', 'List', kTestValues2['flutter.List'])
       ]);
-      expect(preferences.getString('String'), kTestValues2['flutter.String']);
-      expect(preferences.getBool('bool'), kTestValues2['flutter.bool']);
-      expect(preferences.getInt('int'), kTestValues2['flutter.int']);
-      expect(preferences.getDouble('double'), kTestValues2['flutter.double']);
-      expect(preferences.getStringList('List'), kTestValues2['flutter.List']);
+      final all = await preferences.getAll();
+      expect(all['String'], kTestValues2['flutter.String']);
+      expect(all['bool'], kTestValues2['flutter.bool']);
+      expect(all['int'], kTestValues2['flutter.int']);
+      expect(all['double'], kTestValues2['flutter.double']);
+      expect(all['List'], kTestValues2['flutter.List']);
     });
 
     testWidgets('removing', (WidgetTester _) async {
       const String key = 'testKey';
-      await preferences.setString(key, kTestValues['flutter.String']);
-      await preferences.setBool(key, kTestValues['flutter.bool']);
-      await preferences.setInt(key, kTestValues['flutter.int']);
-      await preferences.setDouble(key, kTestValues['flutter.double']);
-      await preferences.setStringList(key, kTestValues['flutter.List']);
+
+      await Future.wait([
+        preferences.setValue('String', key, kTestValues['flutter.String']),
+        preferences.setValue('Bool', key, kTestValues['flutter.bool']),
+        preferences.setValue('Int', key, kTestValues['flutter.int']),
+        preferences.setValue('Double', key, kTestValues['flutter.double']),
+        preferences.setValue('StringList', key, kTestValues['flutter.List'])
+      ]);
       await preferences.remove(key);
-      expect(preferences.get('testKey'), isNull);
+      final all = await preferences.getAll();
+      expect(all['testKey'], isNull);
     });
 
     testWidgets('clearing', (WidgetTester _) async {
-      await preferences.setString('String', kTestValues['flutter.String']);
-      await preferences.setBool('bool', kTestValues['flutter.bool']);
-      await preferences.setInt('int', kTestValues['flutter.int']);
-      await preferences.setDouble('double', kTestValues['flutter.double']);
-      await preferences.setStringList('List', kTestValues['flutter.List']);
+      await Future.wait(<Future<bool>>[
+        preferences.setValue('String', 'String', kTestValues['flutter.String']),
+        preferences.setValue('Bool', 'bool', kTestValues['flutter.bool']),
+        preferences.setValue('Int', 'int', kTestValues['flutter.int']),
+        preferences.setValue('Double', 'double', kTestValues['flutter.double']),
+        preferences.setValue('StringList', 'List', kTestValues['flutter.List'])
+      ]);
       await preferences.clear();
-      expect(preferences.getString('String'), null);
-      expect(preferences.getBool('bool'), null);
-      expect(preferences.getInt('int'), null);
-      expect(preferences.getDouble('double'), null);
-      expect(preferences.getStringList('List'), null);
+      final all = await preferences.getAll();
+      expect(all['String'], null);
+      expect(all['bool'], null);
+      expect(all['int'], null);
+      expect(all['double'], null);
+      expect(all['List'], null);
     });
   });
 }
